bug fix (did not remove indirect relationships when empty)

+++ src/app/controllers/editors/lex-entry-relations-editor/lex-entry-direct-relations/lex-entry-direct-relations-strategy.ts
@@ -71,6 +71,10 @@ export class LexEntryDirectRelationsStrategy implements BaseLexEntityRelationsSt
   }
 
   public removeRelationship(control: FormItem) {
+    if (!control.destinationURI) {
+      return of('');
+    }
+
     return this.lexiconService.deleteRelation(this.lexEntityId, {
       relation: control.relationshipURI,
       value: control.destinationURI,
