+++ src/app/controllers/editors/lex-sense-relations-editor/lex-sense-direct-relations/lex-sense-direct-relations-strategy.ts
@@ -72,6 +72,10 @@ export class LexSenseDirectRelationsStrategy implements BaseLexEntityRelationsSt
   }
 
   public removeRelationship(control: FormItem) {
+    if (!control.destinationURI) {
+      return of('');
+    }
+
     return this.lexiconService.deleteRelation(this.lexEntityId, {
       relation: control.relationshipURI,
       value: control.destinationURI,
